Fixing priority queue and refactoring trees
